requirements.txt
Django==5.0.0
django-allauth>=0.61.0
anthropic>=0.40.0
python-decouple>=3.8
requests>=2.32.0
PyJWT>=2.12.0
cryptography>=46.0.0